removed ctrl, proxy based sigslot
