+++ examples/spa/mod/Row.js
@@ -5,6 +5,6 @@ return {
 	},
 	create:function(deps){
 		this.el.innerHTML=deps.tpl(deps.model)
-		deps.model.desc=deps.model.desc+Date.now()
+		//deps.model.desc=deps.model.desc+Date.now()
 	}
 }
