+++ lib/make.js
@@ -13,7 +13,7 @@ return {
 	run: function(ctx, name, type, params, rawSpec, klass, cb){
 		var _arguments = arguments
 		specMgr.load(ctx, params, rawSpec, function(err, spec){
-			if (err) return cb(err)
+			if (err) return cb(null, rawSpec)
 
 			var cargs = map[type]
 			if (!cargs) return cb(null, spec)
