+++ lib/sigslot.js
@@ -1,78 +1,12 @@
-var specMgr=require('p/specMgr')
 var evts=[]
-var middlewares=[]
 
-function addMW(arr){
-	for(var i=0,a; (a=arr[i]); i++){
-		middlewares.push(specMgr.getValue(a))
+this.update= function(){
+	for(var i=0, l=evts.length, e; i < l; i++){
+		e=evts.shift()
+		dispatch.call(e[0], e[1], e[2])
 	}
 }
-function applyMW(mws,evt,args,cb){
-	if (!mws.length) return cb(evt,args)
-	var mw=mws.shift()
-	var next=function(evt,args){
-		if (!evt) return cb(evt,args)
-		applyMW(mws,evt,args,cb)
-	}
-	if (mw instanceof Function) mw(evt,args,next)
-	else if (mw.run) mw.run(evt,args,next)
-	else applyMW(mws,evt,args,cb)
-}
-function sigslot(self, def){
-	var signals = {}
 
-    ;(self.signals||[]).concat(def||[]).forEach(function(evt){
-		var sender = this
-		signals[evt] = function(){
-			return {
-				args: Array.prototype.slice.call(arguments),
-				sender: sender,
-				evt: evt,
-				mws:[],
-				queue: false,
-				run:run,
-				send: proc,
-				sendNow: procNow
-			}
-		}
-	}, self)
-
-	self.callback.on('*', recv, self)
-
-	return signals
-}
-function remove(signals){
-	var keys = Object.keys(signals)
-	var signal
-	keys.forEach(function(key){
-		signal = signals[key]()
-		signal.args.length = 0
-		signal.mws.length = 0
-		delete signals[key]
-	})
-}
-function run(mw){
-	if (mw instanceof Function)this.mws.push(mw)
-	else if (mw.length) Array.prototype.push.apply(this.mws,mw)
-	return this
-}
-function proc(a, next){
-	var self=this
-	applyMW(this.mws.concat(middlewares),this.evt,this.args,function(evt,args){
-		if (!evt) return console.warn(self.evt,'signal aborted',args)
-		self.evt=evt
-		self.args=args
-		;(next || send).call(self,a)
-	})
-}
-function procNow(a){
-	proc.call(this, a, dispatch)
-}
-function send(a, from){
-	if (this.sender._removed) return
-	this.queue=true
-	evts.push([this, a, from||this.sender])
-}
 function recv(evt, from, params){
 	var func = this.slots[evt]
 	var forward = true
@@ -81,6 +15,12 @@ function recv(evt, from, params){
 	if (from===this) return // prevent trigger twice from extra 'from'
 	if (forward) (params.queue?send:dispatch).call(params, [from,this], this)
 }
+
+function send(a, from){
+	if (this.sender._removed) return
+	evts.push([this, a, from||this.sender])
+}
+
 function dispatch(a, from){
 	if (this.sender._removed) return
 	from=from||this.sender
@@ -99,15 +39,47 @@ function dispatch(a, from){
 	}
 }
 
-this.update= function(){
-	for(var i=0,l=evts.length,e; i<l; i++){
-		e=evts.shift()
-		dispatch.call(e[0], e[1], e[2])
+var Run = {
+	get(target, propKey, receiver){
+		switch(propKey){
+		default: return
+		case 'send':
+			return function(a){
+				send.call(Object.assign(target, {queue: true}), a)
+			}
+		case 'sendNow':
+			return function(a){
+				dispatch.call(target, a)
+			}
+		}
+	}
+}
+
+var Signal = {
+	get(target, propKey, receiver){
+		if (!target.signals.includes(propKey)) return
+		return function(){
+			return new Proxy({
+				sender: target,
+				evt: propKey,
+				args: Array.prototype.slice.call(arguments),
+			}, Run)
+		}
 	}
 }
 
 return {
-	create:sigslot,
-	remove:remove,
-	addMiddleware:addMW
+	create(self, def){
+		var sigslot = Proxy.revocable(self, Signal)
+		self._sigslot = sigslot
+		self.signal = sigslot.proxy
+		self.callback.on('*', recv, self)
+	},
+	remove(self){
+		self.callback.off('*', recv)
+		var ss = self._sigslot
+		if (!ss) return
+		self._sigslot = self.signal = void 0
+		ss.revoke()
+	}
 }
