+++ lib/specMgr.js
@@ -56,6 +56,7 @@ var
 		if (!Array.isArray(spec)) return cb('invalid spec: ' + JSON.stringify(spec))
 		if (spec.length <= idx) return cb(null, deps, params, userData)
 		var s = spec[idx++]
+		if (!s) return load(host, params, spec, idx, deps, cb, userData)
 		var t = s[TYPE]
 		var f
 
@@ -86,8 +87,7 @@ var
 			if (!m) return cb(ERR2.replace('REF', s[VALUE]).replace('RECORD',e), deps, params, userData)
 			deps.push(create(s[ID], t, m))
 			break
-		case 'ctrl':
-		case 'view': // ID[id/path] TYPE[ctrl/view] VALUE[spec] EXTRA[path/path+mixins]
+		case 'view': // ID[id/path] TYPE[view] VALUE[spec] EXTRA[path/path+mixins]
 			f=s[ID]
 			return loadDeps(s[EXTRA]||f, 0, function(err, klass){
 				if (err) return cb(err, deps, params, userData)
