+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "pico-client",
-  "version": "0.6.3",
+  "version": "0.6.4",
   "description": "pico client",
   "main": "create",
   "repository": {
